[WIP] Fix ball being vertically stuck out of range
When bouncing against ball, use test_motion() to know whether I'm going
again toward the paddle. If so, give a horizontal direction to the ball.

Plus, I notice sometimes the ball goes almost vertically toward the
other player.

I wasn't able however to determine whether this bug happened before this
commit or not.

I'm almost certain that if it was, then it's due to the handling of
collision with paddle through signals (this whole branch is the cause).